make this the default definition of HardFault_Handler
